refactor: refactor namespacing pattern, move event drivers to a a Driver folder inside Event folder
