+++ src/modules/Event/Drivers/UIDriver.js
@@ -4,16 +4,16 @@
 
 /**
  * event initialization options.
- *@typedef {Object} eventInit
- *@property {boolean} [eventInit.bubbles=true] - boolean value indicating if event bubbles
- *@property {boolean} [eventInit.cancelable=false] - boolean value indicating if event is
+ *@typedef {Object} UIEventInit
+ *@property {boolean} [UIEventInit.bubbles=true] - boolean value indicating if event bubbles
+ *@property {boolean} [UIEventInit.cancelable=false] - boolean value indicating if event is
  * cancelable
- *@param {WindowProxy} [eventInit.view=null] - identifies the window from which the event was generated.
- *@param {number} [eventInit.detail=0] - value is initialized to a number that is application-specific.
+ *@property {WindowProxy} [UIEventInit.view=null] - identifies the window from which the event was generated.
+ *@property {number} [UIEventInit.detail=0] - value is initialized to a number that is application-specific.
 */
-import {host} from '../Globals.js';
+import {createDOMEvent} from '../../Globals.js';
 import Driver from './Driver.js';
-import Util from '../Util.js';
+import Util from '../../Util.js';
 
 /**
  * ui event driver class
@@ -30,9 +30,20 @@ export default class UIDriver extends Driver {
     }
 
     /**
+     * event init keys
+     *@type {Array}
+    */
+    static get eventInitKeys() {
+        let keys = Driver.eventInitKeys;
+
+        keys.push('view', 'detail');
+        return keys;
+    }
+
+    /**
      * initializes the event according to the UIEvent interface eventInit requirement
      *@param {Object} storeIn - object in which to store initializations
-     *@param {eventInit} getFrom - event initialization objects
+     *@param {UIEventInit} getFrom - event initialization objects
      *@returns {Object}
     */
     static initEvent(storeIn, getFrom) {
@@ -40,17 +51,20 @@ export default class UIDriver extends Driver {
 
         storeIn.view = Util.isObject(getFrom.view)? getFrom.view : null;
         storeIn.detail = Util.isNumber(getFrom.detail)? getFrom.detail : 0;
+
         return storeIn;
     }
 
     /**
      * creates a UIEvent object that can be dispatched to an event target
      *@param {string} type - the event type
-     *@param {eventInit} eventInit - event initialization object
+     *@param {UIEventInit} eventInit - event initialization object
      *@returns {UIEvent}
     */
     static create(type, eventInit) {
-        return new host.UIEvent(type, this.initEvent({}, eventInit));
+        return createDOMEvent(
+            'UIEvent', type, this.initEvent({}, eventInit), this.eventInitKeys
+        );
     }
 
     /**
